namingTpl utils
