+++ app/components/NamingTemplate.js
@@ -7,6 +7,7 @@ import Col from 'react-bootstrap/Col';
 import Button from 'react-bootstrap/Button';
 
 import Handlebars from 'handlebars';
+import { buildTemplateVars } from '../utils/namingTpl';
 
 // import SyntaxHighlighter from 'react-syntax-highlighter';
 // import {
@@ -20,11 +21,7 @@ import Handlebars from 'handlebars';
 //   vs
 // } from 'react-syntax-highlighter/dist/esm/styles/hljs';
 
-import deepFilter from '../utils/deepFilter';
 import TemplateEditor from './TemplateEditor';
-import Airing from '../utils/Airing';
-import { SERIES, PROGRAM, MOVIE, EVENT } from '../constants/app';
-import getConfig from '../utils/config';
 
 const helpers = require('template-helpers')();
 const sanitize = require('sanitize-filename');
@@ -45,70 +42,6 @@ helpers.lPad = (str: string, len: string | number, char: string | number) => {
 };
 
 Handlebars.registerHelper(helpers);
-
-const buildTemplateVars = async (type: number) => {
-  const config = getConfig();
-  const { episodePath, moviePath, eventPath, programPath } = config;
-
-  const typeRe = new RegExp(`${type}`);
-  const recData = await global.RecDb.asyncFindOne({ path: { $regex: typeRe } });
-  const airing = await Airing.create(recData);
-
-  const path = airing.typePath;
-  const showRec = await global.ShowDb.asyncFindOne({ path });
-  if (showRec) recData.show = showRec;
-
-  const globalVars = {
-    EXT: 'mp4',
-    title: airing.title
-  };
-
-  let typeVars = {};
-  switch (type) {
-    case SERIES:
-      typeVars = {
-        episodePath,
-        showTitle: airing.showTitle,
-        seasonNum: airing.seasonNum,
-        episodeNum: airing.episodeNum
-      };
-      break;
-
-    case MOVIE:
-      typeVars = {
-        moviePath
-      };
-      break;
-    case EVENT:
-      typeVars = {
-        eventPath
-      };
-      break;
-
-    case PROGRAM:
-    default:
-      typeVars = { programPath };
-  }
-
-  // let result: Object = {};
-  const result: Object = deepFilter(recData, (value: any, prop: any) => {
-    // prop is an array index or an object key
-    // subject is either an array or an object
-    // console.log(value, prop, subject);
-    if (prop && prop.toString().includes('path')) return false;
-    if (prop && prop.toString().includes('error')) return false;
-    if (prop && prop.toString().includes('warnings')) return false;
-    if (prop && prop.toString().includes('_id')) return false;
-    if (prop && prop.toString().includes('image')) return false;
-    if (prop && prop.toString().includes('Image')) return false;
-    if (prop && prop.toString().includes('user_info')) return false;
-    if (prop && prop.toString().includes('qualifiers')) return false;
-
-    return true;
-  });
-
-  return { ...globalVars, ...typeVars, ...result };
-};
 
 type Props = { label: string, value: string, type: number };
